+++ spec/Pim/Bundle/CatalogBundle/Manager/ProductTemplateMediaManagerSpec.php
@@ -2,21 +2,21 @@
 
 namespace spec\Pim\Bundle\CatalogBundle\Manager;
 
+use Akeneo\Component\FileStorage\Model\FileInterface;
+use Akeneo\Component\FileStorage\RawFile\RawFileStorerInterface;
 use PhpSpec\ObjectBehavior;
-use Pim\Bundle\CatalogBundle\Manager\MediaManager;
-use Pim\Bundle\CatalogBundle\Model\AttributeInterface;
-use Pim\Bundle\CatalogBundle\Model\ProductMediaInterface;
 use Pim\Bundle\CatalogBundle\Model\ProductTemplateInterface;
 use Pim\Bundle\CatalogBundle\Model\ProductValueInterface;
+use Pim\Component\Catalog\FileStorage;
 use Prophecy\Argument;
-use Symfony\Component\HttpFoundation\File\File;
+use Symfony\Component\HttpFoundation\File\UploadedFile;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 
 class ProductTemplateMediaManagerSpec extends ObjectBehavior
 {
-    function let(MediaManager $mediaManager, NormalizerInterface $normalizer)
+    function let(RawFileStorerInterface $fileStorer, NormalizerInterface $normalizer)
     {
-        $this->beConstructedWith($mediaManager, $normalizer);
+        $this->beConstructedWith($fileStorer, $normalizer);
     }
 
     function it_is_initializable()
@@ -25,31 +25,31 @@ class ProductTemplateMediaManagerSpec extends ObjectBehavior
     }
 
     function it_uses_the_media_manager_to_handle_the_media_of_product_templates(
-        $mediaManager,
+        $fileStorer,
         $normalizer,
         ProductTemplateInterface $template,
         ProductValueInterface $imageValue,
-        ProductMediaInterface $imageMedia,
-        File $imageFile,
-        AttributeInterface $image,
+        FileInterface $imageMedia,
         ProductValueInterface $fileValue,
-        ProductMediaInterface $fileMedia
+        FileInterface $fileMedia,
+        FileInterface $fileMediaUploaded
     ) {
+        $pathname = tempnam(sys_get_temp_dir(), 'spec');
+        $uploadedFile = new UploadedFile($pathname, 'uploaded file.txt');
+
         $normalizer->normalize(Argument::cetera())->willReturn([]);
         $template->getValues()->willReturn([$imageValue, $fileValue]);
         $template->setValuesData([])->willReturn($template);
 
         $imageValue->getMedia()->willReturn($imageMedia);
-        $imageMedia->getFile()->willReturn($imageFile);
-        $imageValue->getAttribute()->willReturn($image);
-        $imageValue->getLocale()->willReturn('en_US');
-        $imageValue->getScope()->willReturn('mobile');
-        $image->getCode()->willReturn('main_image');
+        $imageMedia->isRemoved()->willReturn(true);
+        $imageValue->setMedia(null)->shouldBeCalled();
 
         $fileValue->getMedia()->willReturn($fileMedia);
-
-        $mediaManager->handle($imageMedia, Argument::containingString('-main_image-en_US-mobile-'))->shouldBeCalled();
-        $mediaManager->handle($fileMedia, null)->shouldBeCalled();
+        $fileMedia->isRemoved()->willReturn(false);
+        $fileMedia->getUploadedFile()->willReturn($uploadedFile);
+        $fileStorer->store($uploadedFile, FileStorage::CATALOG_STORAGE_ALIAS, true)->willReturn($fileMediaUploaded);
+        $fileValue->setMedia($fileMediaUploaded)->shouldBeCalled();
 
         $this->handleProductTemplateMedia($template);
     }
@@ -60,12 +60,14 @@ class ProductTemplateMediaManagerSpec extends ObjectBehavior
         ProductTemplateInterface $textTemplate,
         ProductValueInterface $imageValue,
         ProductValueInterface $textValue,
-        ProductMediaInterface $imageMedia
+        FileInterface $imageMedia
     ) {
         $normalizer->normalize(Argument::cetera())->willReturn([]);
 
         $imageTemplate->getValues()->willReturn([$imageValue]);
         $imageValue->getMedia()->willReturn($imageMedia);
+        $imageMedia->isRemoved()->willReturn(true);
+        $imageValue->setMedia(null)->shouldBeCalled();
         $imageTemplate->setValuesData([])->shouldBeCalled();
 
         $textTemplate->getValues()->willReturn([$textValue]);
@@ -73,16 +75,5 @@ class ProductTemplateMediaManagerSpec extends ObjectBehavior
 
         $this->handleProductTemplateMedia($imageTemplate);
         $this->handleProductTemplateMedia($textTemplate);
-    }
-
-    function it_generates_the_media_filename_prefix(ProductValueInterface $fileValue, AttributeInterface $file)
-    {
-        $fileValue->getAttribute()->willReturn($file);
-        $file->getCode()->willReturn('file');
-        $fileValue->getLocale()->willReturn('de_DE');
-        $fileValue->getScope()->willReturn('print');
-
-        $prefix = $this->generateFilenamePrefix($fileValue);
-        $prefix->shouldMatch('/-file-de_DE-print-/');
     }
 }