+++ spec/Pim/Bundle/CatalogBundle/Validator/Constraints/FileValidatorSpec.php
@@ -2,8 +2,8 @@
 
 namespace spec\Pim\Bundle\CatalogBundle\Validator\Constraints;
 
+use Akeneo\Component\FileStorage\Model\FileInterface;
 use PhpSpec\ObjectBehavior;
-use Pim\Bundle\CatalogBundle\Model\ProductMediaInterface;
 use Pim\Bundle\CatalogBundle\Validator\Constraints\File;
 use Prophecy\Argument;
 use Symfony\Component\Validator\Context\ExecutionContextInterface;
@@ -21,26 +21,54 @@ class FileValidatorSpec extends ObjectBehavior
         $this->shouldHaveType('Pim\Bundle\CatalogBundle\Validator\Constraints\FileValidator');
     }
 
-    function it_validates_extensions($context, File $constraint)
-    {
+    function it_validates_extensions(
+        $context,
+        File $constraint,
+        FileInterface $file
+    ) {
         $constraint->allowedExtensions = ['gif', 'jpg'];
+        $file->getId()->willReturn(12);
+        $file->getUploadedFile()->willReturn(null);
+        $file->getExtension()->willReturn('jpg');
+        $file->getSize()->willReturn(100);
 
         $context
             ->buildViolation(Argument::any())
             ->shouldNotBeCalled();
 
-        $this->validate(
-            new \SplFileInfo(__DIR__.'/../../../../../../features/Context/fixtures/akeneo.jpg'),
-            $constraint
-        );
+        $this->validate($file, $constraint);
+    }
+
+    function it_validates_size(
+        $context,
+        File $constraint,
+        FileInterface $file
+    ) {
+        $constraint->maxSize = '1M';
+
+        $file->getId()->willReturn(12);
+        $file->getUploadedFile()->willReturn(null);
+        $file->getExtension()->willReturn('jpg');
+        $file->getSize()->willReturn(500);
+
+        $context
+            ->buildViolation(Argument::any())
+            ->shouldNotBeCalled();
+
+        $this->validate($file, $constraint);
     }
 
     function it_does_not_validate_extensions(
         $context,
         File $constraint,
+        FileInterface $file,
         ConstraintViolationBuilderInterface $violation
     ) {
         $constraint->allowedExtensions = ['pdf', 'docx'];
+        $file->getId()->willReturn(12);
+        $file->getUploadedFile()->willReturn(null);
+        $file->getExtension()->willReturn('jpg');
+        $file->getSize()->willReturn(100);
 
         $context
             ->buildViolation(
@@ -50,32 +78,46 @@ class FileValidatorSpec extends ObjectBehavior
             ->shouldBeCalled()
             ->willReturn($violation);
 
-        $this->validate(
-            new \SplFileInfo(__DIR__.'/../../../../../../features/Context/fixtures/akeneo.jpg'),
-            $constraint
-        );
+        $this->validate($file, $constraint);
     }
 
-    function it_validates_extensions_with_product_media(
+    function it_does_not_validate_size(
         $context,
         File $constraint,
-        ProductMediaInterface $productMedia
+        FileInterface $file,
+        ConstraintViolationBuilderInterface $violation
     ) {
-        $constraint->allowedExtensions = ['gif', 'jpg'];
-        $productMedia->getFile()->willReturn(
-            new \SplFileInfo(__DIR__.'/../../../../../../features/Context/fixtures/akeneo.jpg')
-        );
+        $constraint->maxSize = '1M';
+        $file->getId()->willReturn(12);
+        $file->getUploadedFile()->willReturn(null);
+        $file->getExtension()->willReturn('jpg');
+        $file->getSize()->willReturn(1075200);
+        $file->getOriginalFilename()->willReturn('my file.jpg');
 
         $context
-            ->buildViolation(Argument::any())
-            ->shouldNotBeCalled();
+            ->buildViolation($constraint->maxSizeMessage)
+            ->shouldBeCalled()
+            ->willReturn($violation);
+        $violation->setParameter('{{ file }}', Argument::any())->shouldBeCalled()->willReturn($violation);
+        $violation->setParameter('{{ size }}', Argument::any())->shouldBeCalled()->willReturn($violation);
+        $violation->setParameter('{{ limit }}', Argument::any())->shouldBeCalled()->willReturn($violation);
+        $violation->setParameter('{{ suffix }}', Argument::any())->shouldBeCalled()->willReturn($violation);
+        $violation->setCode(Argument::any())->shouldBeCalled()->willReturn($violation);
+        $violation->addViolation()->shouldBeCalled();
 
-        $this->validate($productMedia, $constraint);
+        $this->validate($file, $constraint);
     }
 
-    function it_validates_nullable_value($context, File $constraint)
-    {
+    function it_validates_new_instance_of_files(
+        $context,
+        File $constraint,
+        FileInterface $file
+    ) {
         $constraint->allowedExtensions = ['gif', 'jpg'];
+        $constraint->maxSize = '2M';
+
+        $file->getId()->willReturn(null);
+        $file->getUploadedFile()->willReturn(null);
 
         $context
             ->buildViolation(Argument::any())
@@ -84,17 +126,31 @@ class FileValidatorSpec extends ObjectBehavior
         $this->validate(null, $constraint);
     }
 
-    function it_validates_empty_extensions($context, File $constraint)
+    function it_validates_nullable_value($context, File $constraint)
     {
-        $constraint->allowedExtensions = [];
+        $constraint->allowedExtensions = ['gif', 'jpg'];
+        $constraint->maxSize = '2M';
 
         $context
             ->buildViolation(Argument::any())
             ->shouldNotBeCalled();
 
-        $this->validate(
-            new \SplFileInfo(__DIR__.'/../../../../../../features/Context/fixtures/caterpillar_variant_import.zip'),
-            $constraint
-        );
+        $this->validate(null, $constraint);
+    }
+
+    function it_validates_empty_extension_and_size($context, File $constraint, FileInterface $file)
+    {
+        $constraint->allowedExtensions = [];
+
+        $file->getId()->willReturn(12);
+        $file->getUploadedFile()->willReturn(null);
+        $file->getExtension()->willReturn('jpg');
+        $file->getSize()->willReturn(100);
+
+        $context
+            ->buildViolation(Argument::any())
+            ->shouldNotBeCalled();
+
+        $this->validate($file, $constraint);
     }
 }