+++ src/Akeneo/Component/FileStorage/FileFactoryInterface.php
@@ -15,10 +15,9 @@ interface FileFactoryInterface
 {
     /**
      * @param \SplFileInfo $rawFile     the raw file to create a File with
-     * @param array        $pathInfo    the path infos of the file returned by a PathGeneratorInterface
      * @param string       $destFsAlias the filesystem alias where the file will be stored
      *
      * @return FileInterface
      */
-    public function create(\SplFileInfo $rawFile, array $pathInfo, $destFsAlias);
+    public function createFromRawFile(\SplFileInfo $rawFile, $destFsAlias);
 }