+++ src/Akeneo/Component/FileStorage/RawFile/RawFileFetcher.php
@@ -3,7 +3,6 @@
 namespace Akeneo\Component\FileStorage\RawFile;
 
 use Akeneo\Component\FileStorage\Exception\FileTransferException;
-use Akeneo\Component\FileStorage\Model\FileInterface;
 use League\Flysystem\FilesystemInterface;
 
 /**
@@ -19,23 +18,23 @@ class RawFileFetcher implements RawFileFetcherInterface
     /**
      * {@inheritdoc}
      */
-    public function fetch(FileInterface $file, FilesystemInterface $filesystem)
+    public function fetch(FilesystemInterface $filesystem, $fileKey)
     {
-        if (!$filesystem->has($file->getKey())) {
-            throw new \LogicException(sprintf('The file "%s" is not present on the filesystem.', $file->getKey()));
+        if (!$filesystem->has($fileKey)) {
+            throw new \LogicException(sprintf('The file "%s" is not present on the filesystem.', $fileKey));
         }
 
         $localPathname = tempnam(sys_get_temp_dir(), 'raw_file_fetcher_');
 
-        if (false === $stream = $filesystem->readStream($file->getKey())) {
+        if (false === $stream = $filesystem->readStream($fileKey)) {
             throw new FileTransferException(
-                sprintf('Unable to fetch the file "%s" from the filesystem.', $file->getKey())
+                sprintf('Unable to fetch the file "%s" from the filesystem.', $fileKey)
             );
         }
 
         if (false === file_put_contents($localPathname, $stream)) {
             throw new FileTransferException(
-                sprintf('Unable to fetch the file "%s" from the filesystem.', $file->getKey())
+                sprintf('Unable to fetch the file "%s" from the filesystem.', $fileKey)
             );
         }